feat: update OAuth Service and controller
